refactoring, player update
