+++ src/directions/index.js
@@ -2,4 +2,3 @@ export * from './directions';
 export { default as directionIsAllowed } from './directionIsAllowed';
 export { default as switchDirection } from './switchDirection';
 export { default as directionSwitchMap } from './directionSwitchMap';
-export { default as getNewDirectionFromKeyboard } from './getNewDirectionFromKeyboard';
