+++ src/loop/utils/index.js
@@ -1,2 +1,1 @@
 export { default as moveCamera } from './moveCamera';
-export { default as isInTheMiddle } from './isInTheMiddle';
