+++ src/player/utils/index.js
@@ -1,0 +1,2 @@
+export { default as isInTheMiddle } from './isInTheMiddle';
+export { default as getNewDirectionFromKeyboard } from './getNewDirectionFromKeyboard';
