+++ src/player/utils/isInTheMiddle.js
